Bug 626273 - Correct boundary seek logic.  Change media element IDL floats to doubles per the spec.  Invalidate updated frame after seeking.  Optimize small forward seeks in WebM.  r=cpearce a=roc

+++ content/media/nsBuiltinDecoderStateMachine.h
@@ -154,7 +154,7 @@ public:
     mDecoder->GetMonitor().AssertCurrentThreadIn();
     return mState; 
   }
-  virtual void SetVolume(float aVolume);
+  virtual void SetVolume(double aVolume);
   virtual void Shutdown();
   virtual PRInt64 GetDuration();
   virtual void SetDuration(PRInt64 aDuration);
@@ -164,8 +164,8 @@ public:
 
   virtual nsHTMLMediaElement::NextFrameStatus GetNextFrameStatus();
   virtual void Decode();
-  virtual void Seek(float aTime);
-  virtual float GetCurrentTime();
+  virtual void Seek(double aTime);
+  virtual double GetCurrentTime();
   virtual void ClearPositionChangeFlag();
   virtual void SetSeekable(PRBool aSeekable);
   virtual void UpdatePlaybackPosition(PRInt64 aTime);
@@ -470,7 +470,7 @@ protected:
   // Volume of playback. 0.0 = muted. 1.0 = full volume. Read/Written
   // from the state machine and main threads. Synchronised via decoder
   // monitor.
-  float mVolume;
+  double mVolume;
 
   // PR_TRUE if the media resource can be seeked. Accessed from the state
   // machine and main threads. Synchronised via decoder monitor.
